move assets to each proxy provider and add mayan explorer
