+++ gemstone/src/swapper/proxy/test/mctp_pending.json
@@ -1,0 +1,146 @@
+{
+  "id": "dab51129-0293-4618-842e-c1aafda841de",
+  "trader": "0x867D8D20009B8b875C46dB5904c2CB40380587CD",
+  "traderLedger": null,
+  "sourceTxHash": "0xf151c62aa530d86de11d5c2691b97fc67f0af52e474f9f2abae7fcf794f48594",
+  "orderId": "MCTP_0xf151c62aa530d86de11d5c2691b97fc67f0af52e474f9f2abae7fcf794f48594__0",
+  "sourceTxBlockNo": 28475094,
+  "status": "INITIATED_ON_EVM_MCTP",
+  "transferSequence": "-1",
+  "swapSequence": "6450",
+  "redeemSequence": null,
+  "refundSequence": null,
+  "fulfillSequence": null,
+  "deadline": "2025-04-11T04:05:39.198Z",
+  "sourceChain": "30",
+  "swapChain": "1",
+  "destChain": "23",
+  "destAddress": "0x867d8d20009b8b875c46db5904c2cb40380587cd",
+  "fromTokenAddress": "0x833589fcd6edb6e08f4c7c32d4f71b54bda02913",
+  "fromTokenChain": "30",
+  "fromTokenSymbol": "USDC",
+  "fromAmount": "529.066169",
+  "fromAmount64": "529066169",
+  "toTokenAddress": "0xaf88d065e77c8cc2239327c5edb3a432268e5831",
+  "toTokenChain": "23",
+  "toTokenSymbol": "USDC",
+  "stateAddr": null,
+  "stateNonce": null,
+  "toAmount": "529.033933",
+  "estimateMarketToAmount": null,
+  "transferSignedVaa": null,
+  "swapSignedVaa": null,
+  "redeemSignedVaa": null,
+  "refundSignedVaa": null,
+  "fulfillSignedVaa": null,
+  "savedAt": "2025-04-04T04:05:39.203Z",
+  "initiatedAt": "2025-04-04T04:05:35.000Z",
+  "completedAt": null,
+  "insufficientFees": false,
+  "retries": 0,
+  "swapRelayerFee": "0",
+  "redeemRelayerFee": "0.032236",
+  "refundRelayerFee": "0",
+  "submissionRelayerFee": null,
+  "statusUpdatedAt": "2025-04-04T04:05:39.203Z",
+  "syncRequestedAt": "2025-04-04T04:11:41.849Z",
+  "bridgeFee": "0",
+  "redeemTxHash": null,
+  "refundTxHash": null,
+  "fulfillTxHash": null,
+  "createTxHash": null,
+  "unlockTxHash": null,
+  "auctionAddress": "Awm2zSgzMGTRraAVjRvshqLehy7mJ2Qr3maURDsoDmwi",
+  "unwrapRedeem": false,
+  "unwrapRefund": false,
+  "driverAddress": null,
+  "relayerAddress": null,
+  "auctionMode": null,
+  "stateOpen": null,
+  "cctpReceivedSuiId": null,
+  "cctpMessage": "000000000000000600000003000000000007da720000000000000000000000001682ae6375c4e4a97e4b583bc394c861a46d896200000000000000000000000019330d10d9cc8751218eaf51e8885d058642e08a000000000000000000000000875d6d37ec55c8cf220b9e5080717549d8aa8eca00000000000000000000000000000000833589fcd6edb6e08f4c7c32d4f71b54bda02913000000000000000000000000875d6d37ec55c8cf220b9e5080717549d8aa8eca000000000000000000000000000000000000000000000000000000001f88e8b9000000000000000000000000875d6d37ec55c8cf220b9e5080717549d8aa8eca",
+  "cctpMessageHash": "fb001a71fb1f03324164b634c9c284220d1de37fb98f80ae876ae9b531113ab8",
+  "cctpSolMessageAccount": null,
+  "cctpNonce": "514674",
+  "cctpAttestation": null,
+  "batchFulfilled": null,
+  "gaslessSignature": null,
+  "gaslessPermit": null,
+  "gaslessTx": null,
+  "gasless": null,
+  "initiateContractAddress": "0x1231DEB6f5749EF6cE6943a275A1D3E7486F4EaE",
+  "posAddress": "0x337685fdaB40D39bd02028545a4FfA7D287cC3E2",
+  "mayanAddress": "0x875d6d37EC55c8cF220B9E5080717549d8Aa8EcA",
+  "mayanBps": null,
+  "referrerAddress": null,
+  "referrerBps": null,
+  "unlockRecipient": null,
+  "lockedFundObjectId": null,
+  "forwardedFromAmount": null,
+  "forwardedTokenAddress": null,
+  "forwardedTokenSymbol": null,
+  "penaltyPeriod": null,
+  "baseBond": null,
+  "perBpsBond": null,
+  "gasDrop": "0",
+  "gasDrop64": "0",
+  "payloadId": 1,
+  "orderHash": null,
+  "randomKey": null,
+  "auctionStateAddr": null,
+  "auctionStateNonce": null,
+  "minAmountOut": null,
+  "minAmountOut64": null,
+  "service": "MCTP_BRIDGE",
+  "toTokenPrice": 0.9999797,
+  "fromTokenPrice": 0.9999797,
+  "redeemTxFee": null,
+  "refundTxFee": null,
+  "fromChainNativeTokenPrice": 1791.1836677,
+  "toChainNativeTokenPrice": 1791.1836677,
+  "customPayload": "0x0000000000000000000000000000000000000000000000000000000000000000",
+  "refundAmount": "529.066169",
+  "refundTokenLogoUri": "https://assets.coingecko.com/coins/images/6319/small/usdc.png?1696506694",
+  "refundTokenScannerUrl": "https://basescan.org/token/0x833589fcd6edb6e08f4c7c32d4f71b54bda02913",
+  "refundTokenSymbol": "USDC",
+  "fromTokenLogoUri": "https://assets.coingecko.com/coins/images/6319/small/usdc.png?1696506694",
+  "toTokenLogoUri": "https://assets.coingecko.com/coins/images/6319/small/usdc.png?1696506694",
+  "fromTokenScannerUrl": "https://basescan.org/token/0x833589fcd6edb6e08f4c7c32d4f71b54bda02913",
+  "toTokenScannerUrl": "https://arbiscan.io/token/0xaf88d065e77c8cc2239327c5edb3a432268e5831",
+  "txs": [
+    {
+      "txHash": "0xf151c62aa530d86de11d5c2691b97fc67f0af52e474f9f2abae7fcf794f48594",
+      "goals": [
+        "SEND"
+      ],
+      "scannerUrl": "https://basescan.org/tx/0xf151c62aa530d86de11d5c2691b97fc67f0af52e474f9f2abae7fcf794f48594"
+    }
+  ],
+  "steps": [
+    {
+      "title": "Transfer 529.066169 USDC from base",
+      "status": "COMPLETED",
+      "type": "INFO"
+    },
+    {
+      "title": "Get CCTP attestation",
+      "status": "ACTIVE",
+      "type": "BLOCK_COUNTER",
+      "meta": {
+        "startBlock": 28475094,
+        "wChainId": 30,
+        "minimumConfirmation": 10000
+      }
+    },
+    {
+      "title": "Redeem trade on arbitrum",
+      "status": "PENDING",
+      "type": "INFO"
+    }
+  ],
+  "clientStatus": "INPROGRESS",
+  "refundChain": "23",
+  "refundTokenAddress": "0x833589fcd6edb6e08f4c7c32d4f71b54bda02913",
+  "clientRelayerFeeSuccess": "0.032236",
+  "clientRelayerFeeRefund": "0.032236"
+}
