+++ gemstone/src/swapper/proxy/test/swift_refunded.json
@@ -1,0 +1,181 @@
+{
+  "id": "aa5dea3b-747f-40ed-a319-a3bd27b9f468",
+  "trader": "0x514BCb1F9AAbb904e6106Bd1052B66d2706dBbb7",
+  "traderLedger": null,
+  "sourceTxHash": "0xcb10f22a381a463e748e10013575a502b7b61489435bb8581b16eb9eed3bbc9a",
+  "orderId": "SWIFT_0xa2142d256b229f8583dbc7d93fdd167038da67b4fa6586bd446b0ee7425c431d",
+  "sourceTxBlockNo": 22167246,
+  "status": "ORDER_REFUNDED",
+  "transferSequence": null,
+  "swapSequence": null,
+  "redeemSequence": null,
+  "refundSequence": "145551",
+  "fulfillSequence": null,
+  "deadline": "2025-03-31T15:42:45.000Z",
+  "sourceChain": "2",
+  "swapChain": "1",
+  "destChain": "1",
+  "destAddress": "A21o4asMbFHYadqXdLusT9Bvx9xaC5YV9gcaidjqtdXC",
+  "fromTokenAddress": "0x0000000000000000000000000000000000000000",
+  "fromTokenChain": "2",
+  "fromTokenSymbol": "ETH",
+  "fromAmount": "0.004",
+  "fromAmount64": "4000000000000000",
+  "toTokenAddress": "0x0000000000000000000000000000000000000000",
+  "toTokenChain": "1",
+  "toTokenSymbol": "SOL",
+  "stateAddr": "JAaWGcS4Kqbwyv1fCduMmBd2ab45Ja2twYbFHQgFeQrk",
+  "stateNonce": "255",
+  "toAmount": "0.00334223",
+  "estimateMarketToAmount": null,
+  "transferSignedVaa": null,
+  "swapSignedVaa": null,
+  "redeemSignedVaa": null,
+  "refundSignedVaa": null,
+  "fulfillSignedVaa": null,
+  "savedAt": "2025-03-31T13:38:12.519Z",
+  "initiatedAt": "2025-03-31T13:37:59.000Z",
+  "completedAt": "2025-03-31T15:47:12.516Z",
+  "insufficientFees": false,
+  "retries": 0,
+  "swapRelayerFee": null,
+  "redeemRelayerFee": "0.0002717",
+  "refundRelayerFee": "0.00038607",
+  "submissionRelayerFee": null,
+  "statusUpdatedAt": "2025-03-31T15:47:12.516Z",
+  "syncRequestedAt": "2025-03-31T15:46:11.148Z",
+  "bridgeFee": "0",
+  "redeemTxHash": null,
+  "refundTxHash": "0xfb70313c44c1ce1e476e83c11b6948fc2d970a2c6147a0d3e27afb0e25c250c1",
+  "fulfillTxHash": null,
+  "createTxHash": null,
+  "unlockTxHash": null,
+  "auctionAddress": "9w1D9okTM8xNE7Ntb7LpaAaoLc6LfU9nHFs2h2KTpX1H",
+  "unwrapRedeem": null,
+  "unwrapRefund": null,
+  "driverAddress": null,
+  "relayerAddress": "B88xH3Jmhq4WEaiRno2mYmsxV35MmgSY45ZmQnbL8yft",
+  "auctionMode": 1,
+  "stateOpen": false,
+  "cctpReceivedSuiId": null,
+  "cctpMessage": null,
+  "cctpMessageHash": null,
+  "cctpSolMessageAccount": null,
+  "cctpNonce": null,
+  "cctpAttestation": null,
+  "batchFulfilled": null,
+  "gaslessSignature": null,
+  "gaslessPermit": null,
+  "gaslessTx": null,
+  "gasless": null,
+  "initiateContractAddress": "0x337685fdaB40D39bd02028545a4FfA7D287cC3E2",
+  "posAddress": "0x337685fdaB40D39bd02028545a4FfA7D287cC3E2",
+  "mayanAddress": "0xC38e4e6A15593f908255214653d3D947CA1c2338",
+  "mayanBps": 50,
+  "referrerAddress": "5fmLrs2GuhfDP1B51ziV5Kd1xtAr9rw1jf3aQ4ihZ2gy",
+  "referrerBps": 50,
+  "unlockRecipient": null,
+  "lockedFundObjectId": null,
+  "forwardedFromAmount": null,
+  "forwardedTokenAddress": null,
+  "forwardedTokenSymbol": null,
+  "penaltyPeriod": null,
+  "baseBond": null,
+  "perBpsBond": null,
+  "gasDrop": "0",
+  "gasDrop64": "0",
+  "payloadId": null,
+  "orderHash": "0xa2142d256b229f8583dbc7d93fdd167038da67b4fa6586bd446b0ee7425c431d",
+  "randomKey": "0x8f778bb5467a8ba917b8927a419ddb4dae9a2278323923c7e03008df725eede5",
+  "auctionStateAddr": "Dkr4MQtfyJaM8c5ApNM3Lh2BkaCQDWNK24CJ3JfqXx5N",
+  "auctionStateNonce": "252",
+  "minAmountOut": "0.50132595",
+  "minAmountOut64": "50132595",
+  "service": "SWIFT_SWAP",
+  "toTokenPrice": 124.7854625,
+  "fromTokenPrice": 1810.73022616,
+  "redeemTxFee": null,
+  "refundTxFee": null,
+  "fromChainNativeTokenPrice": 1810.73022616,
+  "toChainNativeTokenPrice": 124.7854625,
+  "customPayload": null,
+  "refundAmount": "0.00334223",
+  "refundTokenLogoUri": "https://statics.mayan.finance/eth.png",
+  "refundTokenScannerUrl": "https://etherscan.io/token/0x0000000000000000000000000000000000000000",
+  "refundTokenSymbol": "ETH",
+  "fromTokenLogoUri": "https://statics.mayan.finance/eth.png",
+  "toTokenLogoUri": "https://statics.mayan.finance/SOL.png",
+  "fromTokenScannerUrl": "https://etherscan.io/token/0x0000000000000000000000000000000000000000",
+  "toTokenScannerUrl": "https://solscan.io/token/0x0000000000000000000000000000000000000000",
+  "txs": [
+    {
+      "txHash": "0xcb10f22a381a463e748e10013575a502b7b61489435bb8581b16eb9eed3bbc9a",
+      "goals": [
+        "SEND"
+      ],
+      "scannerUrl": "https://etherscan.io/tx/0xcb10f22a381a463e748e10013575a502b7b61489435bb8581b16eb9eed3bbc9a"
+    },
+    {
+      "txHash": "3GG6eTGy41EFd7mLbYGjtSQzwySbcM3UyzpA2MSLBrhTQycA65MAcpuGNcTzHZLnEBqoGCESdxv2MhtBcgrr6AY6",
+      "goals": [
+        "CLOSE"
+      ],
+      "scannerUrl": "https://solscan.io/tx/3GG6eTGy41EFd7mLbYGjtSQzwySbcM3UyzpA2MSLBrhTQycA65MAcpuGNcTzHZLnEBqoGCESdxv2MhtBcgrr6AY6"
+    },
+    {
+      "txHash": "2bQ1n8xzay9wowERzfa8ErN7wXA3neTcaB54TP5NVPosA7YvZyi477XGSnsE7qzbN4uYJBGcPPv7eiouBDnPGk1a",
+      "goals": [
+        "CANCEL"
+      ],
+      "scannerUrl": "https://solscan.io/tx/2bQ1n8xzay9wowERzfa8ErN7wXA3neTcaB54TP5NVPosA7YvZyi477XGSnsE7qzbN4uYJBGcPPv7eiouBDnPGk1a"
+    },
+    {
+      "txHash": "4rVNdcPF8a6t4TDog5pZHwWRUzM3sbja68eXCy5CZXZ8XSTapcz3jgJ7PuMHhXaqnVs9nzJgpeEP1mh9rDzBTEV6",
+      "goals": [
+        "REGISTER_ORDER"
+      ],
+      "scannerUrl": "https://solscan.io/tx/4rVNdcPF8a6t4TDog5pZHwWRUzM3sbja68eXCy5CZXZ8XSTapcz3jgJ7PuMHhXaqnVs9nzJgpeEP1mh9rDzBTEV6"
+    },
+    {
+      "txHash": "oZSHVjoUS56rpcm2iNsLaNLtFfk3dq8q4rueZsb5WcVztopqiurtaxFLSC8fDBN3F2o8HndhFp2BVa34NsuXqZY",
+      "goals": [
+        "REGISTER_ORDER"
+      ],
+      "scannerUrl": "https://solscan.io/tx/oZSHVjoUS56rpcm2iNsLaNLtFfk3dq8q4rueZsb5WcVztopqiurtaxFLSC8fDBN3F2o8HndhFp2BVa34NsuXqZY"
+    },
+    {
+      "txHash": "0xfb70313c44c1ce1e476e83c11b6948fc2d970a2c6147a0d3e27afb0e25c250c1",
+      "goals": [
+        "SETTLE"
+      ],
+      "scannerUrl": "https://etherscan.io/tx/0xfb70313c44c1ce1e476e83c11b6948fc2d970a2c6147a0d3e27afb0e25c250c1"
+    }
+  ],
+  "steps": [
+    {
+      "title": "Order created with 0.004 ETH on ethereum",
+      "status": "COMPLETED",
+      "type": "INFO"
+    },
+    {
+      "title": "Fullfill SOL on solana",
+      "status": "FAILED",
+      "type": "INFO"
+    },
+    {
+      "title": "Order canceled",
+      "status": "COMPLETED",
+      "type": "INFO"
+    },
+    {
+      "title": "Order refunded on ethereum",
+      "status": "COMPLETED",
+      "type": "INFO"
+    }
+  ],
+  "clientStatus": "REFUNDED",
+  "refundChain": "2",
+  "refundTokenAddress": "0x0000000000000000000000000000000000000000",
+  "clientRelayerFeeSuccess": null,
+  "clientRelayerFeeRefund": "0.00065777"
+}
